+++ pysui/sui/sui_clients/sync_client.py
@@ -28,7 +28,10 @@ from pysui import (
     SuiConfig,
 )
 from pysui.sui.sui_clients.common import ClientMixin
-from pysui.sui.sui_constants import TESTNET_FAUCET_STATUS_URL
+from pysui.sui.sui_constants import (
+    TESTNET_FAUCET_STATUS_URLV1,
+    DEVNET_FAUCET_STATUS_URLV1,
+)
 
 from pysui.sui.sui_crypto import MultiSig, SuiPublicKey
 from pysui.sui.sui_types.scalars import (
@@ -323,6 +326,7 @@ class SuiClient(ClientMixin):
         version="0.28.0",
         reason="Added fetch_all as currently limited by RPC providers results",
     )
+    @deprecated(version="0.53.0", reason="Transition to GraphQL QueryNode")
     def get_gas(
         self, address: SuiAddress = None, fetch_all: Optional[bool] = False
     ) -> SuiRpcResult:
@@ -343,6 +347,7 @@ class SuiClient(ClientMixin):
         version="0.28.0",
         reason="Added fetch_all as currently limited by RPC providers results",
     )
+    @deprecated(version="0.53.0", reason="Transition to GraphQL QueryNode")
     def get_coin(
         self,
         coin_type: SuiString,
@@ -381,10 +386,10 @@ class SuiClient(ClientMixin):
         """
         for_address = for_address or self.config.active_address
 
-        def _test_faucet_status(task_id: str) -> SuiRpcResult:
+        def _faucet_status(furl: str, task_id: str) -> SuiRpcResult:
             while True:
                 result = self._client.get(
-                    TESTNET_FAUCET_STATUS_URL + task_id,
+                    furl + task_id,
                 ).json()
                 if result["error"]:
                     return SuiRpcResult(False, f"Error on status for task {task_id}")
@@ -407,15 +412,22 @@ class SuiClient(ClientMixin):
 
         try:
             if self.config.faucet_url:
-                result = self._client.post(
+                s1 = self._client.post(
                     self.config.faucet_url,
                     headers=GetObjectsOwnedByAddress(for_address).header,
                     json={"FixedAmountRequest": {"recipient": f"{for_address}"}},
-                ).json()
+                )
+                # If exhausted requests
+                if s1.status_code == 429:
+                    return SuiRpcResult(False, s1.reason_phrase)
+                result = s1.json()
                 if result["error"] is None:
-                    if self.config.environment == "testnet":
-                        return _test_faucet_status(result["task"])
-                    return SuiRpcResult(True, None, FaucetGasRequest.from_dict(result))
+                    faucet_status = (
+                        DEVNET_FAUCET_STATUS_URLV1
+                        if self.config.environment == "devnet"
+                        else TESTNET_FAUCET_STATUS_URLV1
+                    )
+                    return _faucet_status(faucet_status, result["task"])
                 return SuiRpcResult(False, result["error"])
             else:
                 return SuiRpcResult(False, "Faucet not enabled for this configuration.")
@@ -426,6 +438,7 @@ class SuiClient(ClientMixin):
         except TypeError as texc:
             return SuiRpcResult(False, "Type error", vars(texc))
 
+    @deprecated(version="0.53.0", reason="Transition to GraphQL QueryNode")
     def get_object(
         self, identifier: ObjectID, version: SuiInteger = None
     ) -> Union[SuiRpcResult, Exception]:
@@ -448,6 +461,7 @@ class SuiClient(ClientMixin):
         version="0.28.0",
         reason="Added fetch_all as currently limited by RPC providers results",
     )
+    @deprecated(version="0.53.0", reason="Transition to GraphQL QueryNode")
     def get_objects(
         self,
         address: SuiAddress = None,
@@ -492,6 +506,7 @@ class SuiClient(ClientMixin):
         version="0.29.0",
         reason="Handles large identifier list",
     )
+    @deprecated(version="0.53.0", reason="Transition to GraphQL QueryNode")
     def get_objects_for(
         self, identifiers: list[ObjectID]
     ) -> Union[SuiRpcResult, Exception]:
@@ -519,6 +534,7 @@ class SuiClient(ClientMixin):
 
         return result
 
+    @deprecated(version="0.53.0", reason="Transition to GraphQL QueryNode")
     def get_package(self, package_id: ObjectID) -> Union[SuiRpcResult, Exception]:
         """get_package Get details of Sui package.
 
@@ -532,6 +548,7 @@ class SuiClient(ClientMixin):
         result = self.execute(GetPackage(package=package_id))
         return result
 
+    @deprecated(version="0.53.0", reason="Transition to GraphQL QueryNode")
     def get_events(
         self,
         *,