+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- [change](https://github.com/FrankC01/pysui/issues/184) Devnet Faucet V1
+- Deprecating chain data queries on Sync JSON-RPC SuiClient and Async JSON-RPC SuiClient
+
 ### Removed
 
 ## [0.52.0] - 2024-02-12